CoyoteServer/requirements.txt
aiohttp>=3.12.15
pydglab-ws>=1.1.0
qrcode>=8.2
#setuptools>=78.1.1
pillow>=11.3.0
